requirements.txt
rank_bm25==0.2.2
requests==2.31.0
setuptools==69.0.2
google-generativeai==0.5.2
PyPDF2==3.0.1
python-dotenv==1.0.0